Update to hyprland 0.4.0-alpha2
0.3 is not compatible with current versions of hyprland anymore.
So using an alpha version is better than this tool not working at all.

+++ src/main.rs
@@ -31,7 +31,7 @@ enum CursorZoomCommands {
 
 #[derive(clap::Args)]
 struct MoveCurrentWorkspaceToMonitorCommand {
-    id: u8,
+    id: i128,
 }
 
 fn cursor_zoom(command: CursorZoomCommand) -> hyprland::Result<()> {
